Fix -Wshorten-64-to-32 warnings

+++ test/test_fp_manipulation.cpp
@@ -45,7 +45,7 @@ struct fp_manipulation_test
             array_type expected;
             std::transform(input.cbegin(), input.cend(), expected.begin(),
                            [this](const value_type& v)
-                           { return std::ldexp(v, exponent); });
+                           { return std::ldexp(v, static_cast<int>(exponent)); });
             batch_type res = xsimd::ldexp(batch_input(), bexp);
             INFO("ldexp");
             CHECK_BATCH_EQ(res, expected);
